+++ tests/hlsl2glsltest/hlsl2glsltest.cpp
@@ -479,7 +479,7 @@ static std::string GetCompiledShaderText(ShHandle parser)
 		for (int i = 0; i < count; ++i)
 		{
 			char buf[1000];
-			snprintf(buf,1000,"// %s:%s type %d arrsize %d\n", uni[i].name, uni[i].semantic?uni[i].semantic:"<none>", uni[i].type, uni[i].arraySize);
+			snprintf(buf,1000,"// %s:%s type %d arrsize %d\n", uni[i].name, uni[i].semantic?uni[i].semantic:"<none>", uni[i].type, (int)uni[i].arraySize);
 			txt += buf;
 		}
 	}
